+++ changelog.md
@@ -1,5 +1,11 @@
 # ChangeLogs
 
+## `v1.0.3`
+
+- Fixed LanguageProvider
+- Rework the worker
+- Use ace-lua-linter instead of lua-service for lua linting and formatting
+
 ## `v1.0.2`
 
 - Bug fixes